requirements.txt
torch
spaces
gradio==4.42.0
gradio-imageslider
numpy==1.26.4
transformers
accelerate
diffusers
fastapi<0.113.0
opencv-python
devicetorch
pillow
